% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/fitGaussianGraph.R
\name{fitGaussianGraph}
\alias{fitGaussianGraph}
\title{Fitting bridgeless Gaussian graphical model}
\usage{
fitGaussianGraph(X, n, LambdaA, OmegaA, learning_rate,
  convergence_rate = 0)
}
\arguments{
\item{X}{centered data set.}

\item{LambdaA}{adjacency matrix for the directed edges in G.}

\item{OmegaA}{adjacency matrix for the bidirected edges in G.}

\item{learning_rate}{learning rate for the hill-climbing algorithm.}

\item{convergence_rate}{the difference of the objective function between iterations whence we determine the hill-climbing algorithm has converged.}
}
\value{
A list with components
\describe{
  \item{Lambdahat}{a square matrix of the fitted regression coefficients.}
  \item{Omegahat}{the fitted covariance matrix of the error terms.}
  \item{Sigmahat}{the fitted covariance matrix of all observed variables.}
  \item{pvalue}{the p-value of whether the data set X fits the graphical model.}
}
}
\description{
Fits a bridgeless Gaussian graphical model as well as outputting the p-value of whether the data set fits the model.
}
\details{
The p-value should be used as a guide for rejecting models only as we might no longer be working in a concave space at low p-values.
}
